fix: Improve error handling with attached response and fix some typing issues
