+++ asknews_sdk/errors.py
@@ -1,12 +1,24 @@
-from typing import Optional
+from typing import Dict, Optional
+
+from asknews_sdk.response import APIResponse
 
 
 class APIError(Exception):
+    """
+    AskNews API Error
+    """
     code: int = 500000
+    detail: str = "Internal Server Error"
 
-    def __init__(self, detail: str, code: Optional[int] = None) -> None:
+    def __init__(
+        self,
+        response: APIResponse,
+        detail: Optional[str] = None,
+        code: Optional[int] = None
+    ) -> None:
+        self.response = response
         self.code = code or self.code
-        self.detail = detail
+        self.detail = detail or self.detail
 
     def __str__(self) -> str:
         return f"{self.__class__.__name__}: {self.code} - {self.detail}"
@@ -14,35 +26,46 @@ class APIError(Exception):
 
 class BadRequestError(APIError):
     code = 400000
+    detail = "Bad Request"
 
 
 class ResourceNotFoundError(APIError):
     code = 404000
+    detail = "Resource Not Found"
 
 
 class UnauthorizedError(APIError):
     code = 401000
+    detail = "Unauthorized"
 
 
 class ForbiddenError(APIError):
     code = 403000
+    detail = "Forbidden"
 
 
 class MethodNotAllowed(APIError):
     code = 405000
+    detail = "Method Not Allowed"
 
 
 class ValidationError(APIError):
     code = 422000
-    detail: dict  # type: ignore
+    detail: Dict = {}
 
-    def __init__(self, detail: dict, code: Optional[int] = None) -> None:
+    def __init__(
+        self,
+        response: APIResponse,
+        detail: Optional[Dict] = None,
+        code: Optional[int] = None
+    ) -> None:
+        self.response = response
         self.code = code or self.code
-        self.detail = detail
-
+        self.detail = detail or self.detail
 
 class ServiceUnavailableError(APIError):
     code = 503000
+    detail = "Service Unavailable"
 
 
 ErrorMap = {
@@ -61,8 +84,9 @@ ErrorMap = {
 }
 
 
-def raise_from_json(json: dict) -> None:
+def raise_from_response(response: APIResponse) -> None:
+    json: Dict = response.content
     code = json.get("code", json.get("status_code", 500) * 1000)
     detail = json.get("detail")
 
-    raise ErrorMap.get(code, APIError)(detail, code)
+    raise ErrorMap.get(code, APIError)(response, detail, code)
