+++ asknews_sdk/response.py
@@ -1,0 +1,82 @@
+from __future__ import annotations
+
+from typing import Any, Dict
+
+from httpx import Request, Response
+
+from asknews_sdk.types import StreamType
+from asknews_sdk.utils import deserialize
+
+
+class APIResponse:
+    """
+    API Response object returned by the APIClient.
+    """
+    def __init__(
+        self,
+        request: Request,
+        status_code: int,
+        headers: Dict,
+        body: bytes,
+        stream: bool = False,
+    ) -> None:
+        self.request = request
+        self.status_code = status_code
+        self.headers = headers
+        self.body = body
+        self.stream = stream
+        self.content_type = headers.get("content-type", "application/json")
+        self.content = self._deserialize_body() if not self.stream else self.body
+
+    def _deserialize_body(self) -> Any:
+        if self.content_type == "application/octet-stream":
+            return self.body
+        elif self.content_type == "application/json":
+            return deserialize(self.body)
+        elif self.content_type == "text/plain":
+            return self.body.decode("utf-8")
+        else:
+            return self.body
+
+    @classmethod
+    def from_httpx_response(
+        cls,
+        response: Response,
+        stream: bool = False,
+        stream_type: StreamType = "bytes",
+        sync: bool = True,
+    ) -> APIResponse:
+        """
+        Create an APIResponse object from an HTTPX Response object.
+
+        :param response: HTTPX Response object
+        :type response: Response
+        :param stream: Stream response content
+        :type stream: bool
+        :param stream_type: Stream type
+        :type stream_type: StreamType
+        :param sync: Synchronous or asynchronous
+        :type sync: bool
+        :return: APIResponse object
+        :rtype: APIResponse
+        """
+        if stream:
+            match stream_type:
+                case "bytes":
+                    response_body = response.iter_bytes() if sync else response.aiter_bytes()
+                case "lines":
+                    response_body = response.iter_lines() if sync else response.aiter_lines()
+                case "raw":
+                    response_body = response.iter_raw() if sync else response.aiter_raw()
+                case _:
+                    raise ValueError(f"Invalid stream type: {stream_type}")
+        else:
+            response_body = response.content
+
+        return cls(
+            request=response.request,
+            status_code=response.status_code,
+            headers=dict(response.headers.items()),
+            body=response_body,
+            stream=stream,
+        )
