+++ asknews_sdk/types.py
@@ -10,3 +10,11 @@ RequestAuth = Union[
 ]
 
 StreamType: TypeAlias = Literal["bytes", "lines", "raw"]
+
+
+class Sentinel:
+    def __repr__(self):
+        return self.__class__.__name__
+
+
+CLIENT_DEFAULT = Sentinel()
